Text recognition using the vision framework
Still need to filter images in order to feed them into Tesseract, but
this version draws red rectangles on areas of text using the vision
framework from Apple. Need to remove previously detected areas as
UIViews containing the detected text areas are being drawn one on top
of the other.